Implemented basic minterpy descriptors
