+++ mala/__init__.py
@@ -7,7 +7,8 @@ predict LDOS and postprocess LDOS into energies (and forces, soon).
 
 from .version import __version__
 from .common import Parameters, printout, check_modules
-from .descriptors import Bispectrum, Descriptor, AtomicDensity
+from .descriptors import Bispectrum, Descriptor, AtomicDensity, \
+    MinterpyDescriptors
 from .datahandling import DataHandler, DataScaler, DataConverter, Snapshot
 from .network import Network, Tester, Trainer, HyperOpt, \
     HyperOptOptuna, HyperOptNASWOT, HyperOptOAT, Predictor, \
