+++ mala/descriptors/__init__.py
@@ -2,3 +2,4 @@
 from .bispectrum import Bispectrum
 from .atomic_density import AtomicDensity
 from .descriptor import Descriptor
+from .minterpy_descriptors import MinterpyDescriptors
